Prep for 24x24 and 48x48 optimized logos
* Attempts to load specific logo from tokens repo vs tokenlist url
* Splits token definitions from other constants
* Remove trailing slash from LANDING_PAGE and ANALYTICS_PAGE
* Defaults to logo.png until manually updated later
