+++ src/constants/coinLists.ts
@@ -93,7 +93,7 @@ import {
   COOK,
   SKILL,
   TUS
-} from 'src/constants'
+} from 'src/constants/tokens'
 import { WAVAX } from '@pangolindex/sdk'
 
 export const COIN_LISTS = [
