+++ src/hooks/useColor.ts
@@ -3,9 +3,10 @@ import { shade } from 'polished'
 import Vibrant from 'node-vibrant'
 import { hex } from 'wcag-contrast'
 import { Token } from '@pangolindex/sdk'
+import { getTokenLogoURL_24_24 } from '../constants'
 
 async function getColorFromToken(token: Token): Promise<string | null> {
-  const path = `https://raw.githubusercontent.com/pangolindex/tokens/main/assets/${token.address}/logo.png`
+  const path = getTokenLogoURL_24_24(token.address)
 
   return Vibrant.from(path)
     .getPalette()
