+++ src/pages/Airdrop/index.tsx
@@ -14,7 +14,7 @@ import {
 import { useActiveWeb3React } from '../../hooks'
 import Confetti from '../../components/Confetti'
 import { useTokenBalance } from '../../state/wallet/hooks'
-import { UNI, SUSHI } from '../../constants'
+import { UNI, SUSHI } from '../../constants/tokens'
 import { ChainId, JSBI } from '@pangolindex/sdk'
 import { useTranslation } from 'react-i18next'
