+++ src/pages/Beta/Pool/DetailModal/EarnDetail/index.tsx
@@ -7,7 +7,7 @@ import Stat from 'src/components/Stat'
 import { StakingInfo } from 'src/state/stake/hooks'
 import { BIG_INT_ZERO } from 'src/constants'
 import { useMinichefPendingRewards } from 'src/state/stake/hooks'
-import { PNG } from 'src/constants'
+import { PNG } from 'src/constants/tokens'
 import { useActiveWeb3React } from 'src/hooks'
 import ClaimDrawer from '../../ClaimDrawer'
 import WithdrawDrawer from '../../WithdrawDrawer'
