+++ src/pages/Beta/Swap/WatchList/index.tsx
@@ -3,7 +3,7 @@ import { Box, Button } from '@pangolindex/components'
 import { ChainId, Token } from '@pangolindex/sdk'
 import { Plus } from 'react-feather'
 import { ThemeContext } from 'styled-components'
-import { PNG } from 'src/constants'
+import { PNG } from 'src/constants/tokens'
 import { useActiveWeb3React } from 'src/hooks'
 import WatchlistRow from './WatchlistRow'
 import { WatchListRoot, GridContainer, Title, DesktopWatchList, MobileWatchList } from './styleds'
