+++ src/pages/Stake/Manage.tsx
@@ -18,7 +18,8 @@ import { CountUp } from 'use-count-up'
 
 import { wrappedCurrency } from 'src/utils/wrappedCurrency'
 import usePrevious from 'src/hooks/usePrevious'
-import { BIG_INT_ZERO, PNG, ZERO_ADDRESS } from 'src/constants'
+import { BIG_INT_ZERO, ZERO_ADDRESS } from 'src/constants'
+import { PNG } from 'src/constants/tokens'
 import CurrencyLogo from 'src/components/CurrencyLogo'
 import StakingModalSingleSide from 'src/components/earn/StakingModalSingleSide'
 import UnstakingModalSingleSide from 'src/components/earn/UnstakingModalSingleSide'
