+++ src/pages/Vote/VotePage.tsx
@@ -17,7 +17,8 @@ import VoteModal from '../../components/vote/VoteModal'
 import { TokenAmount, JSBI } from '@pangolindex/sdk'
 import { useTokenBalance } from '../../state/wallet/hooks'
 import { useActiveWeb3React } from '../../hooks'
-import { PNG, ZERO_ADDRESS } from '../../constants'
+import { ZERO_ADDRESS } from '../../constants'
+import { PNG } from '../../constants/tokens'
 import { isAddress, getEtherscanLink } from '../../utils'
 import { useTranslation } from 'react-i18next'
