+++ src/pages/Vote/index.tsx
@@ -5,7 +5,8 @@ import { useAllProposalData, ProposalData, useUserVotes, useUserDelegatee } from
 import DelegateModal from '../../components/vote/DelegateModal'
 import { useTokenBalance } from '../../state/wallet/hooks'
 import { useActiveWeb3React } from '../../hooks'
-import { PNG, ZERO_ADDRESS } from '../../constants'
+import { ZERO_ADDRESS } from '../../constants'
+import { PNG } from '../../constants/tokens'
 import { JSBI, TokenAmount, ChainId } from '@pangolindex/sdk'
 import { shortenAddress, getEtherscanLink } from '../../utils'
 import Loader from '../../components/Loader'
