+++ src/state/stake/singleSideConfig.ts
@@ -1,5 +1,5 @@
 import { ChainId, WAVAX } from '@pangolindex/sdk'
-import { OOE, APEIN, ORBS, PNG } from '../../constants'
+import { OOE, APEIN, ORBS, PNG } from '../../constants/tokens'
 import { SingleSideStaking } from './hooks'
 
 export const SINGLE_SIDE_STAKING: { [key: string]: SingleSideStaking } = {
