+++ src/utils/useUSDCPrice.ts
@@ -2,7 +2,7 @@
 
 import { ChainId, Currency, currencyEquals, JSBI, Price, WAVAX } from '@pangolindex/sdk'
 import { useMemo } from 'react'
-import { USDCe } from '../constants'
+import { USDCe } from '../constants/tokens'
 import { PairState, usePairs } from '../data/Reserves'
 import { useActiveWeb3React } from '../hooks'
 import { wrappedCurrency } from './wrappedCurrency'
